Fix: DO-6035 Fix for dev mode after the previous static url changes
The alteration to the base url in the vite config broke the development server. There doesn't appear to be a way to change the base url from the config so it needed to be applied in the dev command inserted into the package.json.

The alternate way that did not work was to leave the dev server serving on / and try to update the static path to "". Doing this seems to default back to /static/ though and setting it as / caused double // in a couple of places that broke other things.