Improvement: DO-6248: Fix Tooltip infinite loop, Remove tippy.js, Migrate Tooltip to floating-ui (#427)
Fixes an issue where clicking too fast in a controlled MultiSelect would cause an infinite loop. The cause was identified to be tippy.js. Since tippy.js is now deprecated and archived, we have migrated Tooltip to floating-ui, which we already use for other components. This was the last remaining tippy.js component, so we can now safely remove it from our deps.

The external API did not change and was reimplemented using floating-ui's hooks.

The only other usage was GetReferenceClientRect tippy.js type, which was replaced with a DOMRect implementation since it automatically calculates the other properties.